requirements.txt
Django == 5.1.3
mysql-connector-python == 9.2.0
python-decouple==3.8
pillow==11.1.0